Fold nexst shema into current but leave mechanism in place for future breaking changes
