+++ m4d/Properties/launchSettings.json
@@ -44,7 +44,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development",
         "SEARCHINDEX": "SongIndexTest",
         "ASPNETCORE_VITE": "true",
-        "SEARCHINDEXVERSION": "2"
+        "SEARCHINDEXVERSION": "3"
       },
       "applicationUrl": "https://localhost:5001;http://localhost:5000"
     }
