+++ m4d/appsettings.json
@@ -13,21 +13,21 @@
   "AppConfig": {
     "Endpoint": "https://music4dance.azconfig.io"
   },
-  "SongIndexProd-1": {
-    "endpoint": "https://music4dance.search.windows.net",
-    "indexname": "songs-prod"
-  },
   "SongIndexProd-2": {
     "endpoint": "https://music4dance.search.windows.net",
     "indexname": "songs-prod-2"
   },
-  "SongIndexTest-1": {
+  "SongIndexProd-3": {
     "endpoint": "https://music4dance.search.windows.net",
-    "indexname": "songs-test"
+    "indexname": "songs-prod-3"
   },
   "SongIndexTest-2": {
     "endpoint": "https://music4dance.search.windows.net",
     "indexname": "songs-test-2"
+  },
+  "SongIndexTest-3": {
+    "endpoint": "https://music4dance.search.windows.net",
+    "indexname": "songs-test-3"
   },
   "SongIndexExperimental": {
     "endpoint": "https://music4dance.search.windows.net",
