+++ m4dModels/SongIndexNext.cs
@@ -1,19 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
-
-using Azure.Search.Documents.Indexes.Models;
-using Azure.Search.Documents.Models;
-
-using DanceLibrary;
-
-namespace m4dModels
+﻿namespace m4dModels
 {
-    public class SongIndexNext : FlatSongIndex
+    // SongIndexNext is now merged into SongIndex.
+    // Retain as a stub for future breaking changes.
+    public class SongIndexNext : SongIndex
     {
-        // For Moq
+        // Stub for future use.
         public SongIndexNext()
         {
         }
@@ -22,303 +13,5 @@ namespace m4dModels
         {
         }
 
-        public override bool IsNext => true;
-
-        #region Index Management
-
-        public override SearchIndex BuildIndex()
-        {
-            // TODO - make fields virtual, then BuildIndex can be the same for all indexes
-            var fields = new List<SearchField>
-            {
-                new(SongIdField, SearchFieldDataType.String) { IsKey = true },
-                new(
-                    AltIdField, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = false, IsSortable = false, IsFilterable = true,
-                    IsFacetable = false
-                },
-                new(Song.TitleField, SearchFieldDataType.String)
-                {
-                    IsSearchable = true, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(Song.ArtistField, SearchFieldDataType.String)
-                {
-                    IsSearchable = true, IsSortable = true, IsFilterable = false,
-                    IsFacetable = false
-                },
-                new(
-                    AlbumsField, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    UsersField, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(CreatedField, SearchFieldDataType.DateTimeOffset)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(ModifiedField, SearchFieldDataType.DateTimeOffset)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(EditedField, SearchFieldDataType.DateTimeOffset)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(Song.TempoField, SearchFieldDataType.Double)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(Song.LengthField, SearchFieldDataType.Int32)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(BeatField, SearchFieldDataType.Double)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(Song.EnergyField, SearchFieldDataType.Double)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(MoodField, SearchFieldDataType.Double)
-                {
-                    IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    Song.PurchaseField, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    Song.DanceTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    GenreTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    TempoTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(
-                    OtherTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                },
-                new(Song.SampleField, SearchFieldDataType.String)
-                {
-                    IsSearchable = false, IsSortable = false, IsFilterable = true,
-                    IsFacetable = false
-                },
-                new(
-                    ServiceIds, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true,
-                    IsFacetable = false
-                },
-                new(
-                    CommentsField, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                {
-                    IsSearchable = true, IsSortable = false, IsFilterable = true,
-                    IsFacetable = false
-                },
-                new(PropertiesField, SearchFieldDataType.String)
-                {
-                    IsSearchable = false, IsSortable = false, IsFilterable = false,
-                    IsFacetable = false
-                },
-                IndexFieldFromDanceId("ALL")
-            };
-
-            var ids = Dances.Instance.AllDanceTypes.Where(t => t.Id != "ALL").Select(t => IndexFieldFromDanceId(t.Id));
-            fields.AddRange(ids);
-
-            return Info.BuildIndex(
-                fields,
-                suggesters: searchSuggesters,
-                scoringProfiles: searchScoringProfiles,
-                defaultScoringProfile: "Default",
-                true
-            );
-        }
-
-        protected override IList<SearchSuggester> searchSuggesters => [
-            new SearchSuggester("songs",
-                Song.TitleField, Song.ArtistField, AlbumsField, Song.DanceTags,
-                    Song.PurchaseField, GenreTags, TempoTags, OtherTags,
-                    $"dance_ALL/{TempoTags}", $"dance_ALL/{StyleTags}", $"dance_ALL/{OtherTags}")
-        ];
-
-        private static SearchField IndexFieldFromDanceId(string id)
-        {
-            return new SearchField(BuildDanceFieldName(id), SearchFieldDataType.Complex)
-            {
-                Fields =
-                {
-                   new SearchField(Votes, SearchFieldDataType.Int32)
-                    {
-                        IsSearchable = false, IsSortable = true, IsFilterable = true, IsFacetable = false
-                    },
-                   new(
-                    TempoTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                    {
-                        IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                    },
-                   new(
-                    StyleTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                    {
-                        IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                    },
-                   new(
-                    OtherTags, SearchFieldDataType.Collection(SearchFieldDataType.String))
-                    {
-                        IsSearchable = true, IsSortable = false, IsFilterable = true, IsFacetable = true
-                    },
-                }
-            };
-        }
-
-        protected override object DocumentFromSong(Song song)
-        {
-            var tagMap = DanceMusicService.DanceStats.TagManager.TagMap;
-
-            // Set up the purchase flags
-            var purchase = string.IsNullOrWhiteSpace(song.Purchase)
-                ? []
-                : song.Purchase.ToCharArray().Where(c => MusicService.GetService(c) != null)
-                    .Select(c => MusicService.GetService(c).Name).ToList();
-            if (song.HasSample)
-            {
-                purchase.Add("Sample");
-            }
-
-            if (song.HasEchoNest)
-            {
-                purchase.Add("EchoNest");
-            }
-
-            if (song.BatchProcessed)
-            {
-                purchase.Add("---");
-            }
-
-            if (song.Purchase != null && song.Purchase.Contains('x', StringComparison.OrdinalIgnoreCase))
-            {
-                Trace.WriteLine($"SongId = {song.SongId}, Purchase = {song.Purchase}");
-            }
-
-            var purchaseIds = song.GetExtendedPurchaseIds();
-
-            // And the tags
-            var ts = new TagSummary(song.TagSummary, tagMap);
-            var genre = ts.GetTagSet("Music");
-            var other = ts.GetTagSet("Other");
-            var tempo = ts.GetTagSet("Tempo");
-
-            var dance = song.TagSummary.GetTagSet("Dance");
-
-            var comments = new List<string>();
-            AccumulateComments(song.Comments, comments);
-
-            foreach (var dr in song.DanceRatings)
-            {
-                var dobj = Dances.Instance.DanceFromId(dr.DanceId);
-                if (dobj is DanceGroup)
-                {
-                    continue;
-                }
-                AccumulateComments(dr.Comments, comments);
-            }
-
-            var users = song.ModifiedBy.Select(
-                m =>
-                    m.UserName.ToLower() +
-                    (m.Like.HasValue ? m.Like.Value ? "|l" : "|h" : string.Empty)).ToArray();
-
-            var altIds = song.GetAltids().ToArray();
-
-            var doc = new SearchDocument
-            {
-                [SongIdField] = song.SongId.ToString(),
-                [AltIdField] = altIds,
-                [Song.TitleField] = song.Title,
-                [Song.ArtistField] = song.Artist,
-                [Song.LengthField] = song.Length,
-                [BeatField] = CleanNumber(song.Danceability),
-                [Song.EnergyField] = CleanNumber(song.Energy),
-                [MoodField] = CleanNumber(song.Valence),
-                [Song.TempoField] = CleanNumber((float?)song.Tempo),
-                [CreatedField] = song.Created,
-                [ModifiedField] = song.Modified,
-                [EditedField] = song.Edited,
-                [Song.SampleField] = song.Sample,
-                [Song.PurchaseField] = purchase.ToArray(),
-                [ServiceIds] = purchaseIds.ToArray(),
-                [AlbumsField] = song.Albums.Select(ad => ad.Name).ToArray(),
-                [UsersField] = users,
-                [Song.DanceTags] = dance.ToArray(),
-                [GenreTags] = genre.ToArray(),
-                [TempoTags] = tempo.ToArray(),
-                [OtherTags] = other.ToArray(),
-                [CommentsField] = comments.ToArray(),
-                [PropertiesField] = SongProperty.Serialize(song.SongProperties, null)
-            };
-
-            var allOther = new HashSet<string>();
-            var allTempo = new HashSet<string>();
-            var allStyle = new HashSet<string>();
-
-            // Set the dance ratings
-            foreach (var dr in song.DanceRatings)
-            {
-                var dobj = Dances.Instance.DanceFromId(dr.DanceId);
-                if (dobj is DanceGroup)
-                {
-                    Trace.WriteLine($"Invalid use of group {dobj.Name} in song {song.SongId}");
-                    continue;
-                }
-                var oneTempo = dr.TagSummary.GetTagSet("Tempo");
-                var oneStyle = dr.TagSummary.GetTagSet("Style");
-                var oneOther = dr.TagSummary.GetTagSet("Other");
-                doc[BuildDanceFieldName(dr.DanceId)] = new Dictionary<string, object>
-                {
-                    { Votes, dr.Weight },
-                    { TempoTags, oneTempo.ToArray() },
-                    { StyleTags, oneStyle.ToArray() },
-                    { OtherTags, oneOther.ToArray() }
-                };
-
-                allOther.UnionWith(oneOther);
-                allTempo.UnionWith(oneTempo);
-                allStyle.UnionWith(oneStyle);
-            }
-
-            var all = song.DanceRatings.Sum(dr => dr.Weight);
-            doc["dance_ALL"] = new Dictionary<string, object>
-            {
-                { Votes, all == 0 ? null : all },
-                { TempoTags, allTempo.ToArray() },
-                { StyleTags, allStyle.ToArray() },
-                { OtherTags, allOther.ToArray() }
-            };                
-
-            return doc;
-        }
-        #endregion
-
-        public override async Task<IEnumerable<Song>> FindArtist(string name,
-            CruftFilter cruft = CruftFilter.NoCruft)
-        {
-            return await FindByField(Song.ArtistField, name, "dance_ALL/Votes desc", cruft);
-        }
     }
 }
